+++ changelog.md
@@ -4,4 +4,11 @@
 
 # 1.0.1
 
+Thanks hiimjasmine00
+
 - Fixed about page
+- Restructured sorting and node traversal
+
+# 1.1.0
+
+- Added option to push purchased items to the end of the shop